Move repositories interfaces from game bundle into component
