+++ src/Badger/Bundle/GameBundle/Unlocker/BadgeUnlocker.php
@@ -5,7 +5,7 @@ namespace Badger\GameBundle\Unlocker;
 use Badger\GameBundle\Entity\BadgeInterface;
 use Badger\GameBundle\Entity\ClaimedBadgeInterface;
 use Badger\GameBundle\Factory\UnlockedBadgeFactoryInterface;
-use Badger\GameBundle\Repository\UnlockedBadgeRepositoryInterface;
+use Badger\Component\Game\Repository\UnlockedBadgeRepositoryInterface;
 use Badger\StorageUtilsBundle\Remover\RemoverInterface;
 use Badger\StorageUtilsBundle\Saver\SaverInterface;
 use Badger\UserBundle\Entity\UserInterface;
