+++ src/Badger/Component/Game/Repository/QuestRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Badger\GameBundle\Repository;
+namespace Badger\Component\Game\Repository;
 
 use Badger\TagBundle\Taggable\TaggableInterface;
 use Badger\UserBundle\Entity\UserInterface;
