+++ src/Badger/Component/Game/Repository/TagSearchableRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Badger\GameBundle\Repository;
+namespace Badger\Component\Game\Repository;
 
 use Badger\TagBundle\Entity\TagInterface;
